removed some errors
Fixes to vec3, removed reliance on old mojang vector library, fixed some issues with the peripheral provider, but it still doesn't work. Fixed DataHelpers and A couple other things. For future reference items and blocks need to be added back to the creative tab. ItemProperties where changed so assigning a creative tab is different.

+++ common/src/main/kotlin/net/spaceeye/someperipherals/integrations/cc/peripherals/WorldScannerPeripheral.kt
@@ -77,7 +77,7 @@ class WorldScannerPeripheral(private val level: Level, private val pos: BlockPos
         val range = SomePeripheralsConfig.SERVER.WORLD_SCANNER_SETTINGS.max_allowed_range
         if (range >= 0 && sqrt(x*x+y*y+z*z) > range) { return makeErrorReturn("Block is outside of range") }
 
-        return makeResult(level.getBlockState(BlockPos(x + pos.x, y + pos.y, z + pos.z)))
+        return makeResult(level.getBlockState(BlockPos((x + pos.x).toInt(), (y + pos.y).toInt(), (z + pos.z).toInt())))
     }
 
     @LuaFunction
